refactor(apis): split endpoints to apis/
